partial apply reusable function
